docs: fix most asyncio DeprecationWarnings
Calling get_event_loop() in Python 3.12 and newer when no event loop
is running is deprecated. The recommended option is to use asyncio.run()
instead.
examples/http_server_asyncio.py is not updated as the example
is quite broken as is so changing it without fixing the existing
problems is not wise.